revise: fallback to master branch
